Port r3921 till r3924: - Move VISIBLE_XTILES from src/ to lib/ivis_common/, so we don't have to hardcore mapview sizes in piedraw.c... - Make INTERFACE_DEPTH/MAX_Z int, since that's what they are used as. - Remove unused defines - Less fog and more shadows
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@3926 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_opengl/piefunc.c
@@ -50,7 +50,7 @@ void pie_DrawViewingWindow(Vector3i *v, UDWORD x1, UDWORD y1, UDWORD x2, UDWORD 
 	pieVrts[0].pos.x = v[1].x;
 	pieVrts[0].pos.y = v[1].y;
 	//cull triangles with off screen points
-	pieVrts[0].pos.z  = (int)INTERFACE_DEPTH;
+	pieVrts[0].pos.z  = INTERFACE_DEPTH;
 
 	pieVrts[0].u = 0;
 	pieVrts[0].v = 0;
